converts tools/lookup component to ts and updates to use api service
