+++ ui/app/components/tools/lookup.hbs
@@ -14,7 +14,7 @@
 {{#if this.lookupData}}
   <div class="box is-fullwidth is-sideless is-paddingless is-marginless">
     {{#each-in this.lookupData as |key value|}}
-      {{#let (if (eq key "creation_ttl") "Creation TTL" (to-label key)) as |label|}}
+      {{#let (if (eq key "creationTtl") "Creation TTL" (to-label key)) as |label|}}
         <InfoTableRow @label={{label}} @value={{value}} />
       {{/let}}
     {{/each-in}}
